Add very rudimentary window resizing
Simply split the windows in halves and assign all windows new frames
whenever a window is added or removed.

One problem is that the newly added window doesn't get moved when
move/resized from the window-added callback. It might need to wait for
the next mainloop iteration before attempting to move it.